Created class for newsAPI calls (going to move it later to pod) and removed unnessasary code in test file class
